test(web): colocate grid.util.test
